requirements.txt
fastapi
fastapi[standard]
jinja2
uvicorn
supabase
python-dotenv